api url refactoring
