+++ src/api/socialProfile.js
@@ -1,9 +1,9 @@
 import { apiClient } from './apiClient';
-import { UserController } from './requestUrls';
+import { ProfileController } from './requestUrls';
 
 export const socialProfile = async () => {
   try {
-    const { data } = await apiClient.get(UserController.socialProfile);
+    const { data } = await apiClient.get(ProfileController.socialProfile);
     return data.response;
   } catch (e) {
     console.log(e);
@@ -13,7 +13,7 @@ export const socialProfile = async () => {
 
 export const updateSocialProfile = async (image_key, email, district_id) => {
   try {
-    const { data } = await apiClient.put(UserController.socialProfile, {
+    const { data } = await apiClient.put(ProfileController.socialProfile, {
       districtId: district_id,
     });
     console.log(data);
